App now fetches a user's following/follower list
In addition, each of the following and followers get their data fetched in order to display their display name, handle, and profile picture in the following/followers list.

Upon clicking on one of these users, you will be taken to their profile and be able to view all of their stats.